Add missing cast for private Client constructor

+++ RabbitMQ.Stream.Client/Client.cs
@@ -161,7 +161,7 @@ namespace RabbitMQ.Stream.Client
             _heartBeatHandler = new HeartBeatHandler(
                 SendHeartBeat,
                 Close,
-                parameters.Heartbeat.TotalSeconds);
+                (int) parameters.Heartbeat.TotalSeconds);
             IsClosed = false;
         }
